- drones apagando incendios

+++ src/main/java/ar/edu/itba/pf/domain/environment/objects/DroneObject.java
@@ -2,9 +2,9 @@ package ar.edu.itba.pf.domain.environment.objects;
 
 public interface DroneObject extends EnvironmentObject {
 
-    double getX();
+    int getX();
 
-    double getY();
+    int getY();
 
     double getRadio();
