+++ src/main/java/ar/edu/itba/pf/domain/environment/objects/FieldObject.java
@@ -2,5 +2,5 @@ package ar.edu.itba.pf.domain.environment.objects;
 
 public interface FieldObject extends EnvironmentObject {
 
-    double getZ();
+    int getZ();
 }
