+++ src/main/java/ar/edu/itba/pf/domain/environment/objects/combustible/TreeCO.java
@@ -4,7 +4,7 @@ public class TreeCO extends CombustibleObject{
 
     private final static double COMBUSTION_PER_TIME = .025;
 
-    public TreeCO(double z, double mass, double height) {
+    public TreeCO(int z, double mass, double height) {
         super(z, mass, COMBUSTION_PER_TIME, height);
     }
